Updated register and login architecture, as well as the form handling.
