+++ src/main/webapp/login.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Login</title>
-    <script type="module" src="./login-register.js"></script>
+    <script type="module" src="./login.js"></script>
 </head>
 <!--A simple login form
 String username = request.getParameter("username");
@@ -13,15 +13,15 @@ String password = request.getParameter("password");
 -->
 <body>
     <h1>Login</h1>
-    <div>
+    <form>
         <label for="username">Username:</label>
         <input type="text" name="username" id="username">
         <br>
         <label for="password">Password:</label>
         <input type="password" name="password" id="password">
         <br>
-        <button onClick="handleLogin();">Login</button>
-    </div>
+        <button type="submit" id="login-button">Login</button>
+    </form>
     <p>Don't have an account? <a href="./register.html">Register</a></p>
 </body>
 </html>
