+++ src/main/webapp/register.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Register</title>
-    <script type="module" src="./login-register.js"></script>
+    <script type="module" src="./register.js"></script>
 </head>
 <!--A simple register form
 String username = request.getParameter("username");
@@ -18,7 +18,7 @@ String lastName = request.getParameter("lastName");
 
     <body>
         <h1>Register</h1>
-        <div>
+        <form>
             <label for="username">Username:</label>
             <input type="text" name="username" id="username">
             <br>
@@ -39,8 +39,8 @@ String lastName = request.getParameter("lastName");
             <br>
             <label for="lastName">Last Name:</label>
             <input type="text" name="lastName" id="lastName">
-            <button onClick="handleRegister();">Register</button>
-        </div>
+            <button type="submit" id="register-button">Register</button>
+        </form>
         <p>Already have an account? <a href="./login.html">Login</a></p>
     </body>
